+++ tests/testthat/test-kobo_data.R
@@ -182,6 +182,10 @@ test_that("with kobo_data and set_names_from_varlabel you can use labels as name
                c("tbl_df", "tbl", "data.frame"))
   expect_equal(class(df_dm), "dm")
   expect_error(set_names_from_varlabel(1L))
+
+  expect_equal(df_simple,
+               kobo_data(uid_simple,
+                         colnames_label = TRUE))
 })
 
 test_that("with kobo_data you can read form with 0 submission", {
@@ -243,4 +247,17 @@ test_that("with kobo_data can restrict the data to the latest form version", {
   raw1 <- kobo_data(uid, all_versions = FALSE)
   raw2 <- kobo_data(uid, all_versions = TRUE)
   expect_equal(identical(raw1, raw2), FALSE)
+})
+
+test_that("kobo_data does not work with non survey assets", {
+  skip_on_cran()
+  url <- Sys.getenv("KOBOTOOLBOX_PROD_URL")
+  token <- Sys.getenv("KOBOTOOLBOX_PROD_TOKEN")
+  skip <-  url == "" & token == ""
+  skip_if(skip,
+          message = "Test server not configured")
+
+  kobo_setup(url = url, token = token)
+  uid <- "aDorbmteGaKUCbcn865Grh"
+  expect_error(kobo_data(uid))
 })